Adding IContentSafetyService interface
